+++ Zed-rate-limiter/src/main/java/com/rateLimiter/controller/RateLimitingController.java
@@ -1,6 +1,7 @@
 package com.rateLimiter.controller;
 
-import com.rateLimiter.model.OtpRequest;  // Import OtpRequest class
+import com.rateLimiter.model.OtpRequest;
+import com.rateLimiter.service.BlackListingService;
 import com.rateLimiter.service.RateLimitingService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -10,20 +11,29 @@ import org.springframework.web.bind.annotation.*;
 import java.util.Map;
 
 @RestController
-@RequestMapping("/otp")
+@RequestMapping("/user")
 public class RateLimitingController {
+
+    @Autowired
+    private BlackListingService blackListingService;
 
     @Autowired
     private RateLimitingService rateLimitingService;
 
-    @PostMapping("/send")
+    @PostMapping("/verify")
     public ResponseEntity<String> sendOtp(@RequestBody OtpRequest otpRequest, @RequestHeader Map<String, String> headers) {
-        // Process the OTP request using the service
 
-        System.out.println("Headers: " + headers);
-        boolean isBlocked = rateLimitingService.processOtpRequest(otpRequest, headers);
+        boolean isBlocked = blackListingService.processOtpRequest(otpRequest, headers);
+
         if (isBlocked) {
             return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Carrier or country is blocked");
+        }
+
+        boolean isRateLimitExceeded = !rateLimitingService.isAllowed(otpRequest.getMobile());
+
+        if (isRateLimitExceeded) {
+            // Return 429 Too Many Requests if the rate limit is exceeded
+            return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).body("Rate limit exceeded. Try again later.");
         }
         return ResponseEntity.ok("OTP Sent");
     }
