+++ Zed-rate-limiter/src/main/resources/application.properties
@@ -1,3 +1,9 @@
 aws.accessKeyId=accessKeyId
 aws.secretKey=secretKey
 aws.region=region
+
+spring.redis.host=host.docker.internal
+spring.redis.port=6379
+
+rate.limit.requests=3
+rate.limit.timewindow=60
